ci: :sparkles: Full notification using web-push added
