+++ public/service-worker.js
@@ -1,7 +1,7 @@
 // This code runs when a push is received, even if your site is closed.
 self.addEventListener('push', function (event) {
 	const data = event.data.json();
-	console.log('[Service Worker] Push Received:', data);
+	// console.log('[Service Worker] Push Received:', data);
 
 	const options = {
 		body: data.body,
