+++ src/routes/private/Maintenance.jsx
@@ -19,7 +19,14 @@ export const MaintenanceRoutes = [
 				path: '/maintenance/issue',
 				element: <Issue />,
 				page_name: 'maintenance__issue',
-				actions: ['create', 'read', 'update', 'delete'],
+				actions: [
+					'create',
+					'read',
+					'update',
+					'delete',
+					'notification',
+					'unsubscribe',
+				],
 			},
 		],
 	},
